% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_centralize_matrix}
\alias{calc_centralize_matrix}
\title{helper10: centralize matrix X by column}
\usage{
calc_centralize_matrix(X)
}
\arguments{
\item{X:}{the genotype likelihood matrix, one column for each snp}
}
\value{
a volumn-centeralized matrix for the given input matrix X
}
\description{
called by \code{\link{sample_bootstrap}} and \code{cal_ScoreV_RVS}
}

