% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EG}
\alias{calc_EG}
\alias{calc_EG_Var}
\title{calc_EG calculates the conditional expected value of the genotype given the genotype likelihoods and frequencies E(Gij|Dij)}
\usage{
calc_EG(M, p, rdv)

calc_EG_Var(M, p, rdv)
}
\arguments{
\item{M:}{genotype likelihoods AA, Aa, aa, matrix nsample by 3 (double), genotype likelihood for one locus in VCF;
uses output from \code{calc_pobs_ndepth} function for simulation data and output from \code{getgenexp} or \code{getMAF} for VCF input}

\item{p:}{genotype frequencies AA, Aa, aa (double) for each SNP, vector of length 3; output from \code{calc_EM} function.}

\item{rdv:}{read depth (vector of integers) for all samples, should be a vector of 1 with length=nsample.}

\item{M:}{genotype likelihoods AA, Aa, aa, matrix sum(rdv) by 3 (double);
#' uses output from \code{calc_pobs_ndepth} function for simulation data and output from \code{getgenexp} or \code{getMAF} for VCF input}

\item{p:}{genotype frequencies AA, Aa, aa (double); output from \code{calc_EM} function.}

\item{rdv:}{read depth (vector of integers) for all samples}
}
\value{
EG: a vector containing conditional expectation values. Size should be equal to length(rdv).

E(Gij|Dij)=sum_(g=0)^2 gP(G_ij=g|D_ij)
P(G_ij=g|D_ij)=P(D_ij|G_ij=g)*P(G_ij=g)/P(D_ij),
where P(D_ij|G_ij=g) (input M) are from the VCF file or \code{calc_pobs_ndepth}, p(G_ij=g) (input p) are from \code{calc_EM}.
all values are rescaled by P(D_ij), (see m below, but it is not probability, so m/sum(m) gives us sth. like prob.)

the variance of E(G_ij|D_ij)

##, var(X)=E(X^2)-E(X)^2, E(X) is calculated in calc_EG
}
\description{
It is called from the \code{generate_seqdata_null} and \code{generate_seqdata_alt} functions.
talk to Andriy, he said in the kk loop, he did it for more general case, we only need rdv=1 for our case.
so I renamed Andriy's get_EG to calc_EG_general, delete the kk loop in calc_EG function.

It is called from the \code{generate_seqdata_alt} functions.
talk to Andriy, he said in the kk loop, he did it for more general case, we only need rdv=1 for our case.
so I renamed Andriy's get_EG2 to \code{calc_EG_var_general}, and delete the kk loop in this function.
}
\section{Functions}{
\itemize{
\item \code{calc_EG_Var}: also, see \code{calc_EG},
}}

