% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_ScoreV_RVS}
\alias{calc_ScoreV_RVS}
\title{helper6: Get Variance of vector S by RVS using robust variance}
\usage{
calc_ScoreV_RVS(X1, X2, P, Y)
}
\arguments{
\item{P}{- genotype frequency for J variants P(G=0), P(G=1) and P(G=2), J by 3 matrix}

\item{X1:}{the expected conditional genotype probability for case}

\item{X2:}{the expected conditional genotype probability for control}

\item{Y:}{the phenotype}
}
\value{
the variance of score statistic vector S by  RVS using robust variance
}
\description{
it is called in \code{test_rare_RVS}
}

