% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{combine_twogeno}
\alias{combine_twogeno}
\title{combine_twogeo is used to combine two sets of ped pairs for two samples}
\usage{
combine_twogeno(file1, file2, keep1 = NA, keep2 = NA)
}
\arguments{
\item{file1,}{the file name for the first sample before the extension. Eg. file1=sample1 if the file names are sample1.ped/map}

\item{file2,}{the file name for the second sample before the extension. Eg. file1=sample2 if the file names are sample2.ped/map}

\item{keep1,}{if only a subset of sample1 is needed in the final analyis, keep1=1, ow, keep1=NA}

\item{keep2,}{if only a subset of sample2 is needed in the final analyis, keep2=1, ow, keep2=NA}
}
\value{
a list include data sets 'common' and 'rare' for common  or rare variants and 'snp.Nomiss' for the SNPs kept in common and rare.
}
\description{
combine_twogeo is used to combine two sets of ped pairs for two samples
}

