% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{common_coord}
\alias{common_coord}
\title{common_coord is to used to get intersect between sets}
\usage{
common_coord(cord_set1, cord_set2)
}
\arguments{
\item{cord_set1:}{the coordinate of the snps in file1}

\item{cord_set2:}{the coordinate of the snps in file2}
}
\value{
com_cord: a list include two vectors, one for the location of common snps in file1 and the other for file2.
}
\description{
common_coord is to used to get intersect between sets
}

