% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{convert}
\alias{convert}
\title{convert two columns of a snp from ped file into additive coding}
\usage{
convert(M, ref)
}
\arguments{
\item{ref,}{the first allele for the snp}

\item{M:}{2 columns of a snp from ped file}
}
\value{
Ad, a vector representing additive coding with length nsample for that snp.
}
\description{
convert two columns of a snp from ped file into additive coding
}

