% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{filter_out}
\alias{filter_out}
\title{filter_out filter out snps by missing rate 1-p}
\usage{
filter_out(M1, M2, p)
}
\arguments{
\item{M1,M2}{the additive coding of genotype for case and controls seperately}

\item{p,}{the keeping rate cut off, only keep snps with missing rate <(1-p) in both samples}
}
\value{
list, M, including row combined genotype  and k-index for whether a snp pass the missingness cut off
}
\description{
It removes all the SNPs that missing rate are smaller than 1-p in either sample.
}

