% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{filter_SNPs}
\alias{filter_SNPs}
\title{Filters SNPs based on specific criteria:}
\usage{
filter_SNPs(vcf_file, case, cont, nsnps = 1000, nread = 300, headnum = 9,
  missing_th = 0.5)
}
\arguments{
\item{vcf_file,}{The address/filename' of the vcf file.}

\item{case,}{The cases location in the whole samples, output from the getIDs function.}

\item{cont,}{The controls location in the whole samples, output from the getIDs function.}

\item{nsnps,}{total number of snps in the vcf file. default=1000}

\item{nread,}{each time the read.table will read nread number of rows in file, default=300}

\item{headnum,}{default =9, the number of columns before each individual's data}

\item{missing_th,}{the missing rate for filter out a snp, default is set=0.5}
}
\value{
The function returns 6 matrices and 2 vectors:

6 matrices:case00, case01, case11, cont00, cont01 and cont11 which contain the genotype likelihoods from VCF

2 vectros: chr and position (Loc) for the snps kept for further analysis.
}
\description{
This function read the whole vcf file, filter out SNPs and return the genotype likelihood in VCF file for case and control separately.
it calls \link{\code{get_Likelihood_vcf}} to convert the vcf format into phred score.
}
\details{
the keeping criteria are: SNPs that PASS filters, do not have duplicates, have only one reference  allele and at most 50% missing rate.
}
\examples{
filterSNPs('1g113low_1g56exomehigh_filtered.hg19.chr11_1000snps.vcf',IDs$cases,IDs$controls,)
}

