% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{get_maf}
\alias{get_maf}
\title{get_maf is used to calculate the MAF}
\usage{
get_maf(M)
}
\arguments{
\item{M}{the additive coding of genotype for the whole sample}
}
\value{
maf based on the non-missing data
}
\description{
get_maf is used to calculate the MAF
}

