% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{get_minor}
\alias{get_minor}
\title{get_minor is used to reset the maf if maf>0.5}
\usage{
get_minor(M)
}
\arguments{
\item{M}{the additive coding of genotype for the whole sample}
}
\value{
maf based on the non-missing data
}
\description{
This is not needed if your genotype coding is based on the Num of minor allele
}

