% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{getMAF}
\alias{getMAF}
\title{Calculates the MAF}
\usage{
getMAF(P, MM, chr, loc, ncase, method = "common")
}
\arguments{
\item{P}{Output from the getgenexp function which contains the probabilities of the genotypes.}

\item{MM}{Output matrix from the getgenexp function.}

\item{ncase,}{number of cases}

\item{method,}{need value of common or rare}
}
\value{
The function returns matrix P which contains the probabilities of the genotypes and
 matrices M1 and M2, which can all be inputed into a test function of the rvs package.
}
\description{
Calculates the MAF and returns matrices which are to be inputed into a test function of the rvs package.
}
\examples{
getMAF(geneexp$P,geneexp$MM)
}

