% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{minor_allele}
\alias{minor_allele}
\title{minor_allele is used to find the minor allele in the whole sample}
\usage{
minor_allele(M1, M2)
}
\arguments{
\item{M1:}{the 2 column of a snp from sample1}

\item{M2:}{the 2 column of a snp from sample2}
}
\value{
minor: the minor allele from the sample1+sample2.
}
\description{
minor_allele is used to find the minor allele in the whole sample
}

