% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tests.R
\name{regScore_rare}
\alias{regScore_rare}
\title{It calls functions \code{calc_ScoreV},\code{calc_ScoreV_RVS},\code{calc_ScoreV_likely},\code{test_CAST}, \code{test_Calpha},\code{sample_bootstrap}}
\usage{
regScore_rare(Y, X, nperm)
}
\arguments{
\item{Y}{- phenotype value, 1-cases and 0-controls}

\item{X}{- matrix of conditional expected values/genotype calls for J variants, it has dimensions n by J, J - number of variants, n - sample size}

\item{nperm}{- number of bootstrap}
}
\value{
two p-values for CAST and C-alpha
}
\description{
It calls functions \code{calc_ScoreV},\code{calc_ScoreV_RVS},\code{calc_ScoreV_likely},\code{test_CAST}, \code{test_Calpha},\code{sample_bootstrap}
}

