% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{subset_ind}
\alias{subset_ind}
\title{subset_ind will subset the cases' or controls' genotype from a bigger dataset, and return the genotype for the wanted individuals.}
\usage{
subset_ind(path1, path2)
}
\arguments{
\item{path1:}{input for genotype files, ped file is prefered}

\item{path2:}{the file including the list of IIDs for individuals to be kept}

\item{file3:}{the file name you want your processed genotype and phenotype to be named.}
}
\value{
a list including- geno: the genotype data with the first ncase rows for cases, genotype data starting from 7th column, ncase: Num. of case
}
\description{
subset_ind will subset the cases' or controls' genotype from a bigger dataset, and return the genotype for the wanted individuals.
}

