% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{subset_snp}
\alias{subset_snp}
\title{subset_snp is to extract the genotype columns for chosen snps from a ped format file.}
\usage{
subset_snp(M, set)
}
\arguments{
\item{M,}{the genotype file in ped format,Note, each snp have two columns in the ped file and the first 6 column need to be included.}

\item{set,}{the return of common_coord, the vector of integers representing the location of the common snps in the two sets.}
}
\value{
Msub, the genotype file for the snps common in two sets only, each snp has two columns.
}
\description{
This can also be done by software plink with option --extact snplist.txt, but snpnames are needed in snplist.txt to use plink.
}

