% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{test_Calpha}
\alias{test_Calpha}
\title{helper8: the quadratic test for nsnp rare variant (Calpha)}
\usage{
test_Calpha(S, Sigma)
}
\arguments{
\item{S:}{vector of length nsnp, the Score for each variant}

\item{Sigma:}{matrix Sigma is the variance matrix of vector S.}

\item{nsnp:}{number of snp}
}
\value{
p-values from Calpha test for J variants (double)
}
\description{
it is called in \code{test_rare_regScore}, \code{test_rare_RVS} and \code{test_rare_RVS_regVar}
}

