% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{vcf_process}
\alias{vcf_process}
\title{the main function to call the VCF file and return the  expected probabilities of the genotypes E(G_ij|D_ij) from VCF file}
\usage{

  vcf_process(file = "example/1g113low_1g56exomehigh_filtered.hg19.chr11_1000snps.vcf",
  caseID_file = "example/caseID.txt", IDline = 128, nhead = 9,
  missing_th = 0.2, nsnp = 1000, nread = 300, MAF = 0.05)
}
\arguments{
\item{file,}{the VCF file}

\item{IDline,}{the row number of your VCF which contains all sample ID}

\item{nhead,}{the column number before the sample ID}

\item{missing_th,}{the missing rate cut off for a SNP to be filtered, if missing>missing_th in case or controls, the SNP will be removed}

\item{nsnp,}{should be the number of SNPs in your VCF file, but if VCF is too large, you can choose only read in the first nsnp SNPs}

\item{MAF,}{the MAF to be used to filter out SNP,}

\item{case_ID_file,}{the caseID files}

\item{index,}{if it is common,  SNPs with maf<MAF will be removed, if it is rare, SNPs with maf>MAF will be removed}
}
\value{
matrix of E(G_ij|D_ij)
}
\description{
the main function to call the VCF file and return the  expected probabilities of the genotypes E(G_ij|D_ij) from VCF file
}

